+++ src/types/MessageResponse.ts
@@ -1,7 +1,23 @@
 import { MessageStatus } from './MessageStatus';
 
-export interface MessageResponse {
+export interface ApiMessageResponse {
   count: number;
+  list: {
+    id: string;
+    points: number;
+    number: string;
+    dateSent: number;
+    submittedNumber: string;
+    status: MessageStatus;
+    idx: string | null;
+    parts?: number;
+  }[];
+  message?: string;
+  length: number;
+  parts?: number;
+}
+
+export interface MessageResponse extends Omit<ApiMessageResponse, 'list'> {
   list: {
     id: string;
     points: number;
@@ -12,7 +28,4 @@ export interface MessageResponse {
     idx: string | null;
     parts?: number;
   }[];
-  message?: string;
-  length: number;
-  parts?: number;
 }