+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- Bug with converting message's `date_sent` to `Date`
+
 ## [2.0.7](https://github.com/smsapi/smsapi-javascript-client/compare/v2.0.6...v2.0.7) - 2023-07-17
 
 ### Changed